api/mask/requirements.txt
Shapely-1.6.4.post2-cp36-cp36m-win_amd64.whl
numpy<1.17
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu==1.12.0
keras==2.2.4
opencv-python
h5py
imgaug
IPython[all]
django
djangorestframework
django-background-tasks
mysqlclient